initial class setup for waste water
